requirements.txt
argon2-cffi==16.3.0
astroid==1.5.3
autopep8==1.3.3
cffi==1.11.2
Django==2.0
django-appconf==1.0.2
django-compressor==2.2
django-enumfields==0.9.0
django-sass-processor==0.5.5
isort==4.2.15
lazy-object-proxy==1.3.1
libsass==0.13.3
mccabe==0.6.1
pycodestyle==2.3.1
pycparser==2.18
pylint==1.7.4
pytz==2017.3
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
wrapt==1.10.11
